Arrange files into OrderTaker and ReportGenerator namespaces
