+++ ToyBlockFactory/ToyBlockFactory/ReportGenerator/IReportParser.cs
@@ -1,4 +1,4 @@
-namespace ToyBlockFactory
+namespace ToyBlockFactory.ReportGenerator
 {
     public interface IReportParser
     {
