+++ ToyBlockFactory/ToyBlockFactory/ReportGenerator/InvoiceReportParser.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 
-namespace ToyBlockFactory
+namespace ToyBlockFactory.ReportGenerator
 {
     public class InvoiceReportParser : IReportParser
     {
