+++ ToyBlockFactory/ToyBlockFactory/ReportGenerator/Report.cs
@@ -1,4 +1,4 @@
-namespace ToyBlockFactory
+namespace ToyBlockFactory.ReportGenerator
 {
     public enum Report
     {
