+++ ToyBlockFactory/ToyBlockFactory/ReportGenerator/TableParser.cs
@@ -2,7 +2,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 
-namespace ToyBlockFactory
+namespace ToyBlockFactory.ReportGenerator
 {
     public static class TableParser
     {
